Use async fixtures
I'm not sure this helped at all because some more of the methods need
to be async as far as I know. In particular the `requests` call. I also
thought this might help with the problem of mongods not terminating but
if anything I am now much more confused about subprocesses...