Fix Python workflow triggers
When @frepond opened PR #21, I wasn't able to figure out how to run the
Python workflow jobs which are required for merging. At first I thought
it was because I couldn't find the button to approve them. But I think
it is actually because the triggers for the job were to restrictive.