Add format check job
Add a GitHub Action workflow job to check formatting of code by
running the formatting script then determining if there is a difference
with git. The reason for doing this is that we use yapf and isort and I
am not sure they are compatible which means we can't use their builtin
checking.

This isn't the only option for solving this problem but it is probably
the easiest. One other option would be to write a formatting wrapper
for other formatting tools. Another option would be to write a new
formatter...